[Setting]#18 - 기본 CoreData 파일 DefaultCoreDataView로 이동
